A5: Finished activity 5
added 2 circles, put all the circle setup code in the setupCircles function.

The draw function tells the program what to do in different states, title, simulation or the endings (love or sadness)

Theres a function for each of endings, a function for title & simulation ontop of a function for move, checking if circles are off screen, or if they touch.